pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aethermem"
version = "1.0.0"
description = "A high-resilience memory continuity protocol for AI Agents"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "AGPL-3.0-or-later"}
authors = [
    {name = "AetherMem Authors", email = "contact@aethermem.dev"}
]
maintainers = [
    {name = "AetherMem Core Team", email = "core@aethermem.dev"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Filesystems",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["ai", "memory", "continuity", "vwl", "filesystem", "persistence", "agent"]
dependencies = [
    "numpy>=1.20.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.2",
    "filelock>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "isort>=5.10.0",
    "mypy>=0.910",
    "pre-commit>=2.17.0",
    "twine>=3.7.0",
    "build>=0.7.0",
]
docs = [
    "sphinx>=4.3.0",
    "sphinx-rtd-theme>=1.0.0",
]
emotional = [
    "scikit-learn>=1.0.0",
    "transformers>=4.15.0",
    "sentence-transformers>=2.2.0",
]
openclaw = [
    "openclaw>=2026.3.0",
]

[project.urls]
Homepage = "https://github.com/kric030214-web/AetherMem"
Documentation = "https://github.com/kric030214-web/AetherMem"
Repository = "https://github.com/kric030214-web/AetherMem"
Changelog = "https://github.com/kric030214-web/AetherMem/releases"
Issues = "https://github.com/kric030214-web/AetherMem/issues"

[project.scripts]
aethermem = "aethermem.cli:main"

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=aethermem --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[project.entry-points."openclaw.skills"]
aethermem = "aethermem.integration.openclaw:register_skill"